refactor: use static typing over dynamic dispatch.
Currently, the visitor used for all the AST validation, rewriting and
code generation has two separate layers of virtual function indirection
(and occasionally makes use of a manually-crafted vtable-like hashtable
keyed by the RTTI typeids for hand-rolled dynamic dispatch). This is
more complex than necessary, and limits the opportunities for the
compiler to optimize these passes by e.g. folding a visitor into only
the things that it actually cares about visiting.

This change is intended to be non-functional, and makes use of the C++
Curiously Recurring Template Pattern (CRTP) to make each visitor
implementation fully non-virtual and self-contained. Further, the types
are simplified slightly in order to eliminate the reliance on RTTI, as
this slows things down and adds additional complexity.  This commit
technically removes the need for RTTI at all, however it is still used
within the context of the semantic analyser. This use could be dropped
however, as these already rely on static type assertions (but this
commit is already out-of-control).

The visitor functionality is also extended to allow for arbitrary
information propagation along the visit path, like what was provided by
the `vtable` wrapper. Previously, visitor implementations were using
class members to effectively propagate information through the stack,
with patterns such as:

```
auto scoped_left_del = accept(left_expr);
Value *left_val = expr_; // Effectively a side channel return.
auto scoped_right_del = accept(right_expr);
Value *right_val = expr_; // Don't reorder anything!
```

This leads to fragile code, since it relies on hidden side effects of
the `accept` function. (The `expr_` member isn't really part of the
state of the visitor, it really is intended to be consumed by the
caller.) It is leads to paths where `expr_` or `expr_deleter_` are
cleared or not cleared incorrectly, which can cause correctness issues.

With the static visitor and CRTP, the return types for the `visit`
function are allowed to be specified directly, and information can be
passed through the stack and joined by the visitor in a more direct way.
This also simplifies lifetime management, and `ScopedExpr` is introduced
to subsume the functionality of `expr_deleter_`.