Support `fn` keyword for external names
Fixes #3177

This change adds `fn` to the list of external names that may be used to
index structures, etc. I had originally considered allowing `fn` (and
`config`) arbitrarily as an identifier without the body of a scope
(esentially extending `ident` to include CONFIG and SUBPROG within the
parser, and limiting `attach_point_def` accordingly), but since there is
already a strong split between internal and external identifiers, I went
with the simpler proposed solution.