Prefer static types over double dynamic dispatch
Currently, the visitor used for all the AST validation, rewriting and
code generation has two separate layers of virtual function indirection
(and occasionally makes use of a manually-crafted vtable-like hashtable
keyed by the RTTI typeids for hand-rolled dynamic dispatch). This is
more complex than necessary, and limits the opportunities for the
compiler to optimize these passes by e.g. folding a visitor into only
the things that it actually cares about visiting.

This change is intended to be non-functional, and makes use of the C++
Curiously Recurring Template Pattern (CRTP) to make each visitor
implementation fully non-virtual and self-contained. This commit does
not quite eliminate the need for RTTI at all, which is left to a
subsequent change just to keep each commit manageable. The `Expression*`
and `Statement*` types now have explicit types that are dispatched,
which will be subsequently eliminated.