Accept ASTContext& for Visitors and pass manager entries
In the future, it will be useful to allow the `Visitor` to rewrite parts of the
AST in a consistent way. This requires access to the `ASTContext` in order to
allocate new nodes. Standardize this as required to instantiate new passes.

This is plumbed through the pass, and allows us to generate remove the use of
`Program` since it is already specified in the `ASTContext`.