Accept typed Program as pass manager entry
Since the type is always `Program`, pass directly as a well-typed
reference. Presumably, the original idea was to support passes at
arbitrary nodes (e.g. print applied to some subtree), and allow for a
`Node*` to be returned for graph rewriting. However, this has not
provide necessary, and the top-level program is not rewritten in any
cases. Therefore, this change can be made in prepation for simplifying
the visitor types and removing the runtime double dispatch.