Finish static type conversion
This change converts the `Expression*` and `Statement*` types directly
into `std::variant` types when constructing the AST. This allows them to
be dispatched directly without relying on RTTI or overloaded classes.

There is a need for some helpers to access members, but changes are kept
to a minimum in these cases.