Use structured visitor for code generation.
This commit changes the `CodegenLLVM` pass to use the structured visitor
return type, rather than storing intermediate expressions in `expr_` and
`expr_deleter_`.

The existing `ScopedExprDeleter` is modified to be `ScopedExpr`, which
carries an optional deleter with it. This should can be moved locally to
extend the lifetime for required types (and the class is not copyable).
This moves a lot of the work of checking `expr_deleter_` and the
associated semantics into the type system itself.

This change raised a few minor issues that are fixed inline. Notably,
the control flow has to introduce some explicit annotations for
unreachable paths `__builtin_unreachable`, and handle some error cases.
A lifetime end for some parameters that were previously missed is
automatically injected, changing a few of the codegen tests by adding a
suitable lifetime end to intermediate or temporary variables.
Undoubtedly these are not the only lifetime issues, but I've strived to
make this change as behavior-neutral as possible.

Some of the tests have also changed in ways due to the subtle different
in walking for unary operators on map. Since there is no distinction
between accept/visit, I've had to push the logic down into the
`unop_int` and `unop_ptr` implementations. This means that no separate
visit happens at (the map is *not* treated as an expression). This omits
a dead load/save the previously resulted from that visit, which is just
a load for a variable, but is a sequence of calls for a map.

You can see this in `pointer_inc_var`:

```
diff --git a/tests/codegen/llvm/pointer_inc_var.ll b/tests/codegen/llvm/pointer_inc_var.ll
index f89cae91..b6134b29 100644
--- a/tests/codegen/llvm/pointer_inc_var.ll
+++ b/tests/codegen/llvm/pointer_inc_var.ll
@@ -20,9 +20,8 @@ entry:
   store i64 0, ptr %"$v", align 8
   store i64 1000, ptr %"$v", align 8
   %1 = load i64, ptr %"$v", align 8
-  %2 = load i64, ptr %"$v", align 8
-  %3 = add i64 %2, 2
-  store i64 %3, ptr %"$v", align 8
+  %2 = add i64 %1, 2
+  store i64 %2, ptr %"$v", align 8
   ret i64 0
 }
```

As well as `pointer_inc_map`:

```
diff --git a/tests/codegen/llvm/pointer_inc_map.ll b/tests/codegen/llvm/pointer_inc_map.ll
index e9b9baea..78bda17d 100644
--- a/tests/codegen/llvm/pointer_inc_map.ll
+++ b/tests/codegen/llvm/pointer_inc_map.ll
@@ -18,8 +18,6 @@ declare i64 @llvm.bpf.pseudo(i64 %0, i64 %1) #0
 define i64 @kprobe_f_1(ptr %0) section "s_kprobe_f_1" !dbg !45 {
 entry:
   %"@_newval" = alloca i64, align 8
-  %lookup_elem_val7 = alloca i64, align 8
-  %"@_key2" = alloca i64, align 8
   %lookup_elem_val = alloca i64, align 8
   %"@_key1" = alloca i64, align 8
   %"@_val" = alloca i64, align 8
@@ -50,32 +48,12 @@ lookup_failure:                                   ; preds = %entry
 lookup_merge:                                     ; preds = %lookup_failure, %lookup_success
   %2 = load i64, ptr %lookup_elem_val, align 8
   call void @llvm.lifetime.end.p0(i64 -1, ptr %lookup_elem_val)
-  call void @llvm.lifetime.end.p0(i64 -1, ptr %"@_key1")
-  call void @llvm.lifetime.start.p0(i64 -1, ptr %"@_key2")
-  store i64 0, ptr %"@_key2", align 8
-  %lookup_elem3 = call ptr inttoptr (i64 1 to ptr)(ptr @AT_, ptr %"@_key2")
-  call void @llvm.lifetime.start.p0(i64 -1, ptr %lookup_elem_val7)
-  %map_lookup_cond8 = icmp ne ptr %lookup_elem3, null
-  br i1 %map_lookup_cond8, label %lookup_success4, label %lookup_failure5
-
-lookup_success4:                                  ; preds = %lookup_merge
-  %3 = load i64, ptr %lookup_elem3, align 8
-  store i64 %3, ptr %lookup_elem_val7, align 8
-  br label %lookup_merge6
-
-lookup_failure5:                                  ; preds = %lookup_merge
-  store i64 0, ptr %lookup_elem_val7, align 8
-  br label %lookup_merge6
-
-lookup_merge6:                                    ; preds = %lookup_failure5, %lookup_success4
-  %4 = load i64, ptr %lookup_elem_val7, align 8
-  call void @llvm.lifetime.end.p0(i64 -1, ptr %lookup_elem_val7)
   call void @llvm.lifetime.start.p0(i64 -1, ptr %"@_newval")
-  %5 = add i64 %4, 2
-  store i64 %5, ptr %"@_newval", align 8
-  %update_elem9 = call i64 inttoptr (i64 2 to ptr)(ptr @AT_, ptr %"@_key2", ptr %"@_newval", i64 0)
+  %3 = add i64 %2, 2
+  store i64 %3, ptr %"@_newval", align 8
+  %update_elem2 = call i64 inttoptr (i64 2 to ptr)(ptr @AT_, ptr %"@_key1", ptr %"@_newval", i64 0)
   call void @llvm.lifetime.end.p0(i64 -1, ptr %"@_newval")
-  call void @llvm.lifetime.end.p0(i64 -1, ptr %"@_key2")
+  call void @llvm.lifetime.end.p0(i64 -1, ptr %"@_key1")
   ret i64 0
 }
```

The final noteworth change is in the processing of `lhist`. The visit
pattern previously involved a doulbe walk of the underlying value
expression, resulting in duplicated blocks. Arguably, this is at least
partly hidden by the need to capture `expr_` separately from the
`expr_deleter_` value, which makes this code harder to parse. The bug
can be seen here:

```
diff --git a/src/ast/passes/codegen_llvm.cpp b/src/ast/passes/codegen_llvm.cpp
index 13ea6af7..278c573d 100644
--- a/src/ast/passes/codegen_llvm.cpp
+++ b/src/ast/passes/codegen_llvm.cpp
     Map &map = *call.map;
-    auto scoped_del = accept(call.vargs.front());

     // prepare arguments
     auto *value_arg = call.vargs.at(0);
     auto *min_arg = call.vargs.at(1);
     auto *max_arg = call.vargs.at(2);
     auto *step_arg = call.vargs.at(3);
-    Value *value, *min, *max, *step;
-    auto scoped_del_value_arg = accept(value_arg);
-    value = expr_;
-    auto scoped_del_min_arg = accept(min_arg);
-    min = expr_;
-    auto scoped_del_max_arg = accept(max_arg);
-    max = expr_;
-    auto scoped_del_step_arg = accept(step_arg);
-    step = expr_;
+    auto scoped_value_arg = visit(value_arg);
+    auto scoped_min_arg = visit(min_arg);
+    auto scoped_max_arg = visit(max_arg);
+    auto scoped_step_arg = visit(step_arg);
```

The new bytecode shows that this double value is now eliminated:

```
diff --git a/tests/codegen/llvm/call_lhist.ll b/tests/codegen/llvm/call_lhist.ll
index ae457b62..b0872c7c 100644
--- a/tests/codegen/llvm/call_lhist.ll
+++ b/tests/codegen/llvm/call_lhist.ll
@@ -23,11 +23,8 @@ entry:
   %get_pid_tgid = call i64 inttoptr (i64 14 to ptr)()
   %1 = lshr i64 %get_pid_tgid, 32
   %pid = trunc i64 %1 to i32
-  %get_pid_tgid1 = call i64 inttoptr (i64 14 to ptr)()
-  %2 = lshr i64 %get_pid_tgid1, 32
-  %pid2 = trunc i64 %2 to i32
-  %3 = zext i32 %pid2 to i64
-  %linear = call i64 @linear(i64 %3, i64 0, i64 100, i64 1)
+  %2 = zext i32 %pid to i64
+  %linear = call i64 @linear(i64 %2, i64 0, i64 100, i64 1)
   call void @llvm.lifetime.start.p0(i64 -1, ptr %"@x_key")
   store i64 %linear, ptr %"@x_key", align 8
   %lookup_elem = call ptr inttoptr (i64 1 to ptr)(ptr @AT_x, ptr %"@x_key")
```

This is a follow-on from #3699.