Unambiguously adopt zero value semantics
Rather than relying on semantic analysis to detect cases where variables are
unassigned, we can simplify the code by simply specifying that variables are
always initialized with their respect "zero values", that is the value
specified by an in-memory representation of all zero.

In fact, this is already the case today, both for implementation and
documentation.

For the implementation, during code generation variables of all sizes are
zeroed. We do not make any attempt to optimize for the case where there is a
different initializer, and the generated bytecode often looks like `store 0`,
followed by `store <x>`. This is perfectly reasonable, and we can defer such
analysis and optimization to LLVM. Doing a complete analysis for whether a
variable is assigned can become arbitrarily complex, and in all likelihood we
would need to generate this pattern to avoid undefined behavior (which I
suspect the BPF verified would reject anyways).

Just to provide an example of this analysis being complex, it is already
trivial to escape the existing analysis by simply having the variable defined
in the first branch of an `if` statement:
```
BEGIN {
   let $x: int32;
   let $y: int32;
   $y = 1;
   if ($y == 0) {
     $x = 1;
   } else {
     if ($x == 0) {
        print("x is 0");
     } else {
        print("x is 1");
     }
   }
   exit();
}
```

For the documentation, the man page already states:
```
If no assignment is specified variables will initialize to 0.
```

This change merely drops the associated semantic analyzer checks and metadata,
as well as the test against that specific error.