Simplify max nodes check
This simplifies the max nodes check, which was added for fuzzing.

First, the node visitor can be replaced by a trivial `node_count`
check on the ASTContext, which will directly determine the number
of nodes allocated for the AST without requiring a full walk.

Second, since this check is already implemented (along with the
required environment variable), there is no need to do this with
the fuzzing parse separately -- parsing will already fail.

This allows us to eliminate *both* the visitors (NodeCounter and
CallbackVisitor) and simplify the pass plumbing.