simpify fuzzing infrastructure
This change returns the fuzzing infrastructure to a building state
(from not building), and dramatically simplifies the overall setup.

First, documentation regarding `libfuzzer` is removed, as this is
two generations out of date. (Libfuzzer become Centipede, which in
turn became fuzztest. There's a legitimate use case for fuzztest,
but these are unit tests. A sample is added for this new style of
test for tracepoint parsing, simply as a reference.

Second, the dedicated fuzzing binary is removing and instead the
actual main program can be built with relevant santizing. The
existing `-testmode=codegen` is used to run a fuzz test, and the
existing ability to read an input file is also used. This reduces
duplication and redundancy.

Third, the `#ifdef FUZZ` statements are removed from the code.
All of these statements are either a) never reached with the changes
above, and b) easily mitigated by turning some other knob (such as
redirecting stderr during fuzzing).