cleanup: unify comment style
The codebase current has three different comment styles. In addition to
the C++ style, two C-styles exist. One is the standard multi-line style:

```
/*
 * ...
 */
```

But also a style that omits the leading '*':

```
/**
   ...
 */
```

In a visitor change, the existing comment style was preserved but in a
review the ask was made to change the style for consistency. To do so,
pull forward all the comments in the code base to a single consistent
style, and document this requirement in the coding style docs.

This change uses the C++ style, because it is the most dominant style
currently (accounting for the majority of comments).