ast: restructure the pass manager
Currently the pass manager does little other than run a sequence of
functions in order. This has led to a few problems:

1) The `BPFtrace` object has logically morphed in a kind of "god object",
   where passes set and mutate global state, which is difficult to
   encapsulate, track or mock.

2) Because of (1), tests for individual passes depend on the implicit
   behavior from all previous passes. Therefore, the same pass plumbing
   is replicated in many places.

Finding myself needing to stash some data resulting from passes
(deserialized bitcode from imports), I decided to take a pass at this
problem rather than adding another field to the AST.

This change introduces a pass context and pass manager which operate
using a `PassContext`. Each pass can require some set of known inputs in
addition to the current `ASTContext`, which will be automatically
extracted from the `PassContext`. Furthermore, each pass produces some
set of known outputs, which will automatically be injected into to the
`PassContext` when the pass is finished running.

Attempting to add a pass which does not have the required inputs
produced by an earlier pass will generate an error at add time, not at
runtime. Attempting to set an output which has already been set by a
previous pass will also generate an error.

The core idea for this change is that the intermediate state required
(probe attachment info, resources, etc.) will be manifest as these
objects which live in the context. Known passes can produce these
objects, which will then be consumed subsequently.

This makes it easier to design tests around specific passes: rather than
needing to recreate the known universe, you need only manually inject
the specific bits of context that you care about, and validate that the
intended outputs are successfully produced.

Note that this change still leaves the AST as a mutable structure. It
would be good to allow for incremental transformations there as well
that could be statically validated, but this is a much broader task.
Implementing a simple static storage for passes provides an opportunity
to refactor and simplify a lot of the existing plumbing.

Signed-off-by: Adin Scannell <amscanne@meta.com>