Remove unnecessary class; `std::string` is enough
A `std::string` can easily serve the purposes required here, and there
are already two different constructors. `c_str` will always be
null-terminated, and the `std::string` constructor will copy in state to
ensure that the lifetime of the object isn't bound to external APIs.  A
const reference to the string can be returned to avoid any copies.

Signed-off-by: Adin Scannell <amscanne@meta.com>