cleanup: remove `is_map` and `is_variable`
These are inferrable by checking the dynamic type of the node, which is
used elsewhere. In the future we may consider moving to non-RTTI typing,
but this is orthogonal to eliminating this manual tagging. This cleanup
is being done ahead of more invasive surgery to the pipelines, in order
to minimize unnecessary churn.

Signed-off-by: Adin Scannell <amscanne@meta.com>