ast: support automatic scalar maps
Instead of relying on the `NoneType` for maps, we can support
automatically printing maps as scalars when they are exactly one entry
in size. In the future, this will be move the generation of default keys
up from code generation and effectively "desugar" map functions.

This commit preserves the existing `NoneType` behavior, but allows for
automatic detection of scalar maps, and propagates this information via
the `RequiredResources`. In the future, all maps will have a type after
being desugared, and this detection mechanism will be the only way that
scalar maps work.

Signed-off-by: Adin Scannell <amscanne@meta.com>