ast: allow `replace` to operate on `Expression*` and `Statement*` types
The visitor is starting to get a bit complex, but we can keep using the
existing `replace` pattern for now. This extends it to ensure that it
works on cross-type nodes (e.g. `Expression*` and `Statement*`), which
will allow us to replace some of those node for desugaring.

Signed-off-by: Adin Scannell <amscanne@meta.com>

stack-info: PR: https://github.com/bpftrace/bpftrace/pull/3983, branch: user/amscanne/map_pipelines/6