config: refactor and simplify configuration plumbing
This change started as a way to remove the `StackMode` processing from
the lexer and parser, in order to simplify the literal mechanisms.
Unfortunately, it became quickly apparent that the AST types that these
gave rise to were somewhat tied in to the configuration itself. In order
to achieve this simpliciation, this change refactors and rewrites the
`Config` object.

Instead of a enum class of keys, the `Config` is changed to be a
plain-old C object. This allows us to more throughly document each
member, and reference the configuration is an easily understood and
type-checked way. This also makes the defaults incredibly obvious.

The parsing of different types is delegated to extensible `ConfigParser`
classes, that can be declared for any type that you wish to add to the
configuration. The mapping from configuration "keys" (human strings) is
owned by `config.cpp`, which automatically defines parsers for each of
the associated fields. Amazingly, this is the only place that some
amount of template magic occurs, in order to automatically construct the
appropriate parser based on the field type.

This change does drop some functionality:

* The "ConfigSource" parts are dropped, since they were never referenced
  beyond checking if a configuration could be overriden. There is a much
  simpler way to do this: load the environment variables and override
  configuration values *after* they have been parsed from the script.

* The potential to allow arbitrary expressions as configuration values
  is removed. Arguably, this could have been useful in the future for
  doing evaluation, but this was not done yet. This will be reintroduced
  with proper literal plumbing, coming shortly.

On of the upsides of this change is that parsing is unified for each
type, and we can accept a wider variety of options. For example, in the
configuration block we treat identifiers as strings, so quoting is
necessary only for values containing a space or other characters.
Similarly, values like `true` and `false` can be accepted for boolean
values, in addition to `0` and `1`.

Signed-off-by: Adin Scannell <amscanne@meta.com>

stack-info: PR: https://github.com/bpftrace/bpftrace/pull/3982, branch: user/amscanne/map_pipelines/5