refactor: drop `is_literal` and add folding pass
In order to reduce the complexity inside the more complex passes (the
semantic analyser and code generation), add a folding pass for literals
and constants. Note that this is not intended to do the work that LLVM
will do in the backend, rather it is merely folding constants that must
be interpreted by semantic analysis.

A classic example is `str($1)`, which can be folded into a suitable
`String` node. This eliminates the need for special cases in semantic
analysis and the `is_in_str` field within the AST.

Signed-off-by: Adin Scannell <amscanne@meta.com>

stack-info: PR: https://github.com/bpftrace/bpftrace/pull/3984, branch: user/amscanne/map_pipelines/7