refactor: move to static types
This change makes AST node consistent, and adopts static types rather
than relying on RTTI.

The major motivation for this is to move away from the `Expression`
inheritence model, whereby instances reach into the `Expression` base
class and access the `SizedType` directly. Instead, a layer of
abstraction is added here to allow individual node to provide access to
their type information via a returned reference. Instead of being
specifically a `SizedType` object, this reference can be an abstract
type of some kind, e.g. originating outside of our own type system.

Consider a future pass which examines `Call` nodes, and replaces them
with `ExternalCall` nodes based on the resolve imports. The types for
those new `ExternalCall` nodes should be BTF types sourced from the
impored type information. Currently we would need to figure out a way to
convert every type exhaustively. While this is done currently within the
`BTF` class, it is all driven by the `ClangParser`. Reinvoking the
parser again and again in order to introduce these types to our unified
type universe is not ideal: we should be able to operate exclusively and
directly on the type information contained in the external objects. Over
time, our internal implementations will leverage these mechanisms,
allowing us to remove the special cases from `SizedType`.

With this change, individual node are still typed, but they have their
own field, appropriately named. Direct access to these fields are
encouraged, and this is where type inference will operate. This reduces
ambiguity, as node without fundamental type information (e.g. a
PositionalParameter, a Block, etc.) need not store an invalid value.
They can implement a `type` function which simply references to the
appropriate sub-type.

Note that the `type` method returns immutable references, which means
that type inference and code generation can only mutate types on node
that it understands, which is a desirable safety property. This is on
top of the additional safety properties that no more `static_cast`s or
`dynamic_cast`s are permitted, reducing the possibility of subtle memory
errors.

Most of this change itself is boilerplate in two forms:
* Making AST node consistent in structure and moving the `type`.
* Changing all call-sites to use this new consistent structure.

There should be no semantic change.

Signed-off-by: Adin Scannell <amscanne@meta.com>