codegen: use `ScopedValue` instead of `ScopedExpr`
This should generally be a no-op change. It is using the new class,
which has slightly different semantics in place of the existing one. The
different semantics mean that lifetimes aren't extended quite as much in
some cases, but this should not be an issue. Looking at the existing
cases, the lifetimes were being extended as a very conservative measure.

Signed-off-by: Adin Scannell <amscanne@meta.com>