semantics: introduce `typeof` and `fail` operators
To allow for type-specific behavior in macros and other parts of the
standard library, we need a way to do parameteric behavior based on
types. Rather than overloading, if we introduce a `typeof` operator and
a type comparison expression, then we can simply prune invalid branches
during semantic analysis. These branches can correspond to expanded
macros or anything else.

As a trivial example, consider:

```
macro is_signed($x) {
  let $r;
  if (typeof($x) == typeof(0xffffffffffffffff)) {
    $r = false;
  } else {
    $r = true;
  }
  $r
}

begin {
  if (is_signed(-1)) {
    print("-1 number is signed")
  } else {
    print("-1 number isn't signed")
  }
  if (is_signed(0xffffffffffffffff)) {
    print("big number number is signed")
  } else {
    print("big number number isn't signed")
  }
}
```

To prevent compilation when types are not supported/invalid a `fail`
function is added that serves the same role as a `static_assert`.
Because of folding in the semantic analyser, this can now be used in the
same way. In fact, we can trivially implement `static_assert` as a
macro:

```
macro static_assert($v, $msg) {
  if (typeof($v) != typeof(true)) {
    fail("unable to evaluate compile-time assert");
  } else if (!$v) {
    fail($msg);
  }
}
```

Signed-off-by: Adin Scannell <amscanne@meta.com>

stack-info: PR: https://github.com/bpftrace/bpftrace/pull/4401, branch: user/amscanne/add_typeof/1