return_path: inject returns when needed
This ensures that probes are well-formed with a suitable return already
injected. This eliminates that path from codegen, and allows subsequent
passes to reason about the returns that are available.

In practice, this will be used by the recursion check pass to inject
code that releases the recursion check lock on any return.

This change handles loops and other terminators for basic blocks,
ensuring that we no longer have blocks with no predescessors or deadcode
blocks. Codegen is also no longer responsible for injecting synthetic
branches or returns.

Signed-off-by: Adin Scannell <amscanne@meta.com>

stack-info: PR: https://github.com/bpftrace/bpftrace/pull/4587, branch: user/amscanne/recursion/1