field_analyser: simplify type resolution
The field analyser has many gaps related to assignment of types to
variables and other intermediates. It effectively requires full type
inference to close these gaps. Instead, just fully resolve any explicit
types used in the AST (such as casts, etc.) by recursive evaluation. If
this types are not available, they will be injected into the BTF set for
evaluation by `ClangParser`.

The only real consequence to this is that the types may be more resolved
than they were before, because we will include definitions for fields
that were previously elided. Therefore, the tests must be updated in
these cases to accept more completely-defined tests.

Signed-off-by: Adin Scannell <amscanne@meta.com>