test: add structured AST matchers
Structured AST matchers will allow us to test parsing and transformation
against something that is not the string representation of the AST. The
problem with the string representation of the AST is that it is overly
ensitive to small changes -- we can't specify exactly what parts of the
AST that we care about.

These structured matchers are still optimized for explicitly
specification, but they allow for more flexibility and variability. For
example, you can care that an expression is wrapped in an `if` but then
choose not to match on the branches of that `if`.

Signed-off-by: Adin Scannell <amscanne@meta.com>

stack-info: PR: https://github.com/bpftrace/bpftrace/pull/4674, branch: user/amscanne/gendoc/5